# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoSRE DevOps Agent API",
    "version": "1.0.0",
    "description": "API documentation for the AutoSRE DevOps Agent"
  },
  "servers": [
    {
      "url": "http://localhost:4002"
    }
  ],
  "paths": {
...
    "/api/agent/approve": {
      "post": {
        "summary": "Approve or reject a pending agent action",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Pending action ID"
                  },
                  "approved": {
                    "type": "boolean",
                    "description": "Approval status"
                  }
                },
                "required": ["id", "approved"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval processed successfully"
          },
          "400": {
            "description": "No such pending action or already resolved"
          }
        }
      }
    }
  }
}